requirements.txt
aiogram>=3.10.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
gspread>=6.1.0
oauth2client>=4.1.0
openai>=1.30.0
google-generativeai>=0.7.0
google-genai>=0.1.0
